move pattern_repository out of repositories folder
