+++ lib/app/app_router.dart
@@ -1,7 +1,7 @@
 import 'package:equatable/equatable.dart';
 import 'package:flutter/material.dart';
 
-import '../repositories/pattern_repository/pattern_repository.dart';
+import '../pattern_repository/pattern_repository.dart';
 import '../screens/attributions/attributions.dart';
 import '../screens/home/home.dart';
 import '../screens/search_results/search_results.dart';
