+++ lib/pattern_repository/core/cartesian_product_iterable.dart
