+++ lib/pattern_repository/core/compare.dart
