+++ lib/pattern_repository/core/config_map.dart
