+++ lib/pattern_repository/core/core.dart
