+++ lib/pattern_repository/engine/prechac_pattern.dart
