+++ lib/pattern_repository/engine/prechac_throw_constraint.dart
