+++ lib/pattern_repository/models/pattern.dart
