+++ lib/pattern_repository/models/pattern_constraint.dart
