+++ lib/pattern_repository/models/patternable.dart
