+++ lib/pattern_repository/models/search_parameters.dart
