+++ lib/pattern_repository/models/throw_constraint.dart
