+++ lib/pattern_repository/models/throwable.dart
